refactor: ima screen pass one
